Restore headers back to library location.
