+++ example_code/src/uint128_example/uint128_example.vcxproj.filters
@@ -20,41 +20,6 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <None Include="cjm_string.inl">
-      <Filter>Header Files</Filter>
-    </None>
-    <None Include="numerics.inl">
-      <Filter>Header Files</Filter>
-    </None>
-    <None Include="numerics_configuration.inl">
-      <Filter>Header Files</Filter>
-    </None>
-    <None Include="uint128.inl">
-      <Filter>Header Files</Filter>
-    </None>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="cjm_numeric_concepts.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="cjm_string.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="istream_utils.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="numerics.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="numerics_configuration.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="uint128.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="cjm_configuration.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\..\visual_studio_environment_test\src\visual_studio_environment_test_app\test_program.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
