+++ src/cjm/numerics/cjm_numeric_concepts.hpp
@@ -15,6 +15,8 @@
 #include <optional>
 #include <compare>
 #include <climits>
+#include <cjm/numerics/cjm_configuration.hpp>
+#include <cjm/numerics/numerics_configuration.hpp>
 
 namespace cjm::numerics::concepts
 {
